* 發布第2.0版；         * 當窗口未設置圖符名時，設置圖符名爲窗口標題；         * 增加編譯選項-w，並修復出現警告之處；         * 在運行命令框輸入不完整的命令後按回車，其結果改爲執行第一個匹配的命令；         * 修正手冊頁部分過時的描述，修改部分內容使描述更清晰；         * 修改標題欄按鈕文字，以便於本地化之後能顯示；         * 實現國際化與本地化。

+++ src/layout.c
@@ -52,7 +52,7 @@ static void set_full_layout(WM *wm)
 
 static void set_preview_layout(WM *wm)
 {
-    int n=get_clients_n(wm), i=n-1, rows, cols, w, h,
+    long n=get_clients_n(wm), i=n-1, rows, cols, w, h,
         sw=wm->screen_width, sh=wm->screen_height, ch=(sh-wm->taskbar->h);
     if(n == 0)
         return;
@@ -185,13 +185,13 @@ void update_taskbar_buttons(WM *wm)
 bool is_main_sec_gap(WM *wm, int x)
 {
     Desktop *d=DESKTOP(wm);
-    unsigned int w=wm->screen_width*(1-d->main_area_ratio-d->fixed_area_ratio);
+    long w=wm->screen_width*(1-d->main_area_ratio-d->fixed_area_ratio);
     return (get_typed_clients_n(wm, SECOND_AREA) && x>=w-wm->cfg->win_gap && x<w);
 }
 
 bool is_main_fix_gap(WM *wm, int x)
 {
-    unsigned int w=wm->screen_width*(1-DESKTOP(wm)->fixed_area_ratio);
+    long w=wm->screen_width*(1-DESKTOP(wm)->fixed_area_ratio);
     return (get_typed_clients_n(wm, FIXED_AREA) && x>=w && x<w+wm->cfg->win_gap);
 }
